security cleanup, convertor improvement, product type enum added

+++ src/main/java/com/yayla/secondhand/secondhandbackend/manager/ProductManager.java
@@ -1,7 +1,6 @@
 package com.yayla.secondhand.secondhandbackend.manager;
 
-import com.yayla.secondhand.secondhandbackend.convertor.product.ProductRequestToVoConvertor;
-import com.yayla.secondhand.secondhandbackend.convertor.product.ProductUpdateRequestToVoConvertor;
+import com.yayla.secondhand.secondhandbackend.convertor.product.ProductConvertor;
 import com.yayla.secondhand.secondhandbackend.model.dto.ProductDto;
 import com.yayla.secondhand.secondhandbackend.model.request.ProductCreateRequest;
 import com.yayla.secondhand.secondhandbackend.model.request.ProductUpdateRequest;
@@ -10,8 +9,10 @@ import com.yayla.secondhand.secondhandbackend.model.response.ProductResponse;
 import com.yayla.secondhand.secondhandbackend.model.vo.ProductCreateVo;
 import com.yayla.secondhand.secondhandbackend.model.vo.ProductUpdateVo;
 import com.yayla.secondhand.secondhandbackend.service.ProductService;
+import com.yayla.secondhand.secondhandbackend.service.SessionInfoService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 @Slf4j
@@ -20,8 +21,8 @@ import org.springframework.stereotype.Service;
 public class ProductManager {
 
     private final ProductService productService;
-    private final ProductRequestToVoConvertor productRequestToVoConvertor;
-    private final ProductUpdateRequestToVoConvertor productUpdateRequestToVoConvertor;
+    private final ProductConvertor productConvertor;
+    private final SessionInfoService sessionInfoService;
 
     public ProductResponse fetchProduct(Long productId) {
         log.info("Product fetch has started. productId : {}", productId);
@@ -31,20 +32,25 @@ public class ProductManager {
 
     public ProductResponse createProduct(ProductCreateRequest productCreateRequest) {
         log.info("Product creation has started. productCreateRequest : {}", productCreateRequest.toString());
-        ProductCreateVo productCreateVo = productRequestToVoConvertor.convert(productCreateRequest);
+        Long currentAccountId = sessionInfoService.currentAccountId();
+        ProductCreateVo productCreateVo = productConvertor.convert(productCreateRequest, currentAccountId);
         ProductDto productDto = productService.createProduct(productCreateVo);
         return mapResponse(productDto);
     }
 
     public ProductResponse updateProduct(ProductUpdateRequest productUpdateRequest){
         log.info("Product update has started. productUpdateRequest : {}", productUpdateRequest.toString());
-        ProductUpdateVo productUpdateVo = productUpdateRequestToVoConvertor.convert(productUpdateRequest);
+        Long currentAccountId = sessionInfoService.currentAccountId();
+        validateAccess(productUpdateRequest.getProductId(), currentAccountId);
+        ProductUpdateVo productUpdateVo = productConvertor.convert(productUpdateRequest);
         ProductDto productDto = productService.updateProduct(productUpdateVo);
         return mapResponse(productDto);
     }
 
     public BaseResponse deleteProduct(Long productId) {
         log.info("Product delete has started. productId: {}", productId);
+        Long currentAccountId = sessionInfoService.currentAccountId();
+        validateAccess(productId, currentAccountId);
         productService.deleteProduct(productId);
         return new BaseResponse();
     }
@@ -53,6 +59,13 @@ public class ProductManager {
         ProductResponse productResponse = new ProductResponse();
         productResponse.setProductDto(productDto);
         return productResponse;
+    }
+
+    private void validateAccess(Long productId, Long currentAccountId) {
+        ProductDto productDto = productService.fetchProduct(productId);
+        if(!productDto.getOwnerId().equals(currentAccountId)) {
+            throw new AccessDeniedException("You are not authorized to modify this product");
+        }
     }
 
 }
